backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
python-dotenv>=1.0.0
redis>=5.0.0
google-generativeai>=0.3.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
